requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
python-dotenv==1.0.1
bcrypt==4.2.1
email-validator==2.2.0
pyjwt==2.10.1
fastapi-mail==1.4.1
